pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nutrakinetics-studio"
version = "0.1.0"
description = "Streamlit-first workspace for supplement PBPK and NAD systems modeling."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.40",
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.13",
    "plotly>=5.24",
    "pydantic>=2.8",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
